Comply with ACME spec
ACME spec says that Reply-Nonce is provided only for successful POST
requests or if the nounce is empty or wrong. netproc.c relied on server
sending the Reply-Nonce even for GET requests. This patch changes
behavior to always send POST, first time with intentionally wrong nonce
to be sure that we get the Reply-Nonce back.